[refactor] 중복 saveTest api util 함수 제거
