+++ src/utils/asyncUtils.js
@@ -1,7 +1,6 @@
 import cookie from "react-cookies";
 import { call, put } from "redux-saga/effects";
 import UserAPI from "../api/userAPI";
-import MakingAPI from "../api/makingAPI";
 
 export const [INIT, LOADING, SUCCESS, ERROR] = [
   "init",
@@ -106,9 +105,4 @@ export const reducerUtils = {
     },
     status: ERROR,
   }),
-};
-
-export const saveTest = async (params) => {
-  const { status } = await MakingAPI.saveTest(params);
-  return status;
 };
